Tighten up support for trailing commas in array literals.
The initial support for trailing commas in array literals had a bug
when `deserialize_struct()` was used in conjunction with
`#[serde(deny_unknown_fields)]`. I noticed this because there was a
test case in `test_parse_enum_errors()` that I expected to fail
(because it assumed trailing commas were disallowed), yet it was
still passing (i.e., the trailing comma error was still being thrown).

It appeared that when the number of fields to be deserialized was
known, instead invoking `next_element_seed()` until it returned
`Ok(None)`, serde would stop invoking it after it saw the last
required field and move on to the `end_seq()` call.

Before this change, `end_seq()` would fail because the `SeqAccess`
would not have consumed the final trailing comma. This changes the
implementation of `SeqAccess` so that it is responsible for consuming
the comma associated with a value before returning the value.

Overall, I think this implementation of
`SeqAccess::next_element_seed()` is cleaner than the original.