Rename serde_jsonrc to serde_json_lenient.
The previous owner of serde_jsonrc does not have a need for
serde_jsonrc any longer, so after discussion, the best option
seems to be to make a new crate independently maintained.

The name 'serde_json_lenient' has been chosen because it
hopefully accurately describes the purpose of this crate:
serde_json, but being more lenient to some JSON format quirks.