Do not discard the byte after a lone surrogate when replacing
Given "\ud800X", containing an invalid lone high surrogate, the `X` is
peeked at to see if it is a `\`. If it is not, it should not be consumed
(discarded) before returning REPLACEMENT CHARACTER. However, it must be
consumed if an error will be returned to the caller, to produce the
correct error location.