Improve performance of escape character checking
Prior to this change, the escapes function in read.rs constructs and
returns-by-copy an array of bools for each character processed
inside a string by SliceRead.
This change constructs both of the possible arrays that escapes can
return once, and then changes escapes to return references to them.

On citm_catalog.json (a fairly representative benchmarking input with
a mix of types), this change reduces parsing time by about 15%.
On twitter.json (a string-heavy benchmark), this change reduces
parsing time by 37%.