Correctly handle escape sequences after un-paired high surrogate escapes
Surrogate pairs should be written `\uXXXX\uYYYY`, with U+XXXX a valid
high surrogate. The case where the ``` in that input is present but the
`u` is absent was not handled correctly when `validate` is true and
invalid characters are being replaced: once the replacement is
performed, the backslash-escape was not interpreted, instead just
skipping the `\`.