Rename `read::error` to `read::error_no_replace`
The `error` function exists upstream, while the `error_or_replace` does
not. This means that when upstream adds a reference teo `error` where we
would like to use `error_or_replace` instead, it's easy to miss the
change, if it is not covered by tests.

With this PR, such a new reference to `error` will fail as no such
function is defined, and we will need to decide which function to call.